requirements.txt
Flask==3.0.0
PyMySQL==1.1.0
python-dotenv==1.0.0
matplotlib==3.8.0
pandas==2.1.0
plotly==5.17.0
cryptography==41.0.0